crawler/requirements.txt
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
pydantic>=2.7
requests>=2.31.0
